// src/modules/merchandising/reports/work-order/accessories-report-with-po/accessories-with-po-type.ts
export type IAccessoriesReportWithPo = {
  EXTRA_PERCENT: string;
  REVICE_REASON: string;
  REVICE_DATE: string;
  REVICE_NO: number;
  ID: number;
  WORK_ORDER_NO: string;
  MST_ID: string;
  COMPANY_ID: string;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  COMPANY_EMAIL: string;
  COMPANY_WEB: string;
  COMPANY_REMARKS: string;
  COMPANY_CONTACT: string;
  SUPPLIER_ID: string;
  SUPPLIER_NAME: string;
  SUPPLIER_ADDRESS: string;
  ISSUE_DATE: string;
  IS_APPROVE: string;
  APPROVE_DATE: string;
  REMARKS: string;
  RECEIVE_STORE_ID: string;
  STORE_PERSONNEL: string;
  MOBILE_NO: string;
  RCV_STORE: string;
  DELIVERY_DATE: string;
  WO_SUBJECT: string;
  ATTENTION: string;
  SUPPLIER_MOBILE_NO: string;
  TOTAL_AMOUNT: string;
  BUYER_ID: string;
  PO_ID: string;
  BUYER_NAME: string;
  PO_NO: string;
  STYLE_ID: string;
  STYLENO: string;
  STYLENAME: string;
  MATERIAL_ID: string;
  MTL_NAME: string;
  GMT_COLOR_ID: string;
  GMT_SIZE_ID: string;
  MTL_COLOR_ID: string;
  MTL_SIZE_ID: string;
  TOTAL_BOOKING_QTY: number;
  UOM_ID: string;
  UOM: string;
  SUPPLIER_RATE_PER_PCS: number;
  SUPPLIER_RATE_PER_DZN: number;
  TOTAL_SUP_AMOUNT: number;
  MTL_DELIVERY_DATE: string;
  WORK_ORDER_QTY: number;
  GMT_COLOR_NAME: string;
  GMT_COLOR_CODE: string;
  MTL_COLOR_NAME: string;
  GMT_SIZE_NAME: string;
  MTL_SIZE_NAME: string;
  SORTINGNO: string;
  FABRIC_DIA: string;
  PCS_PER_CTN: number;
  NET_WEIGHT: number;
  GROSS_WEIGHT: number;
  COMPOSITION: string;
  FACTORY_CODE: string;
  NOTE: string;
  ADDITIONAL_IN_QTY: number;
  ADDITIONAL_IN_PERCENTAGE: number;
  ADDITIONAL_QTY: number;
  GMT_QTY: number;
  DESCRIPTION: string;
  APPROVED_BY: string;
  APPROVED_BY_DESG: string;
  PREPARED_BY: string;
  PREPARED_BY_DESG: string;
  PREPARED_BY_MOBILENO: string;
  ORDER_PLACE_MONTH: string;
  CURRENCY: string;
  BUYING_HOUSE: string;
  DESTINATION: string;
  LOT: string;
  SOURCE: string;
  SPECIAL_INSTRUCTION: string;
  PAYMENT_TERMS: string;
  SESSIONNO: string;
  FABRIC_QUALITY: string;
  BRAND_NAME: string;
  PRODUCTFAMILY: string;
  MTL_DESCRIPTION_2: string;
  SUPPLIER_EMAIL: string;
  MTL_COLOR_NAME2: string;
  MTL_COLOR_NAME_2: string;
  BODY_PER_PCS: string;
  SUB_PO: string;
  SHIP_DATE: Date;
  CONSUMPTION_PER_UNIT: number
};